# api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/orders":{"get":{"operationId":"OrderController_getAll","summary":"get orders by status","parameters":[{"name":"status","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"]},"post":{"operationId":"OrderController_post","summary":"create order","parameters":[{"name":"authorization","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDTO"}}}},"responses":{"201":{"description":""}},"tags":["orders"]}},"/api/v1/orders/unfinished":{"get":{"operationId":"OrderController_getAllUnfinished","summary":"get orders unfinished","parameters":[],"responses":{"200":{"description":""}},"tags":["orders"]}},"/api/v1/orders/{id}/status-pagamento":{"get":{"operationId":"OrderController_getPaymentStatus","summary":"get order payment status","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"]}},"/api/v1/orders/checkout":{"put":{"operationId":"OrderController_put","summary":"finish de order","parameters":[{"name":"orderId","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["orders"]}},"/api/v1/orders/{id}/status":{"patch":{"operationId":"OrderController_updateStatus","summary":"update status order","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusOrderDTO"}}}},"responses":{"200":{"description":""}},"tags":["orders"]}},"/api/v1/clients":{"get":{"operationId":"ClientController_getByCpf","summary":"get client by cpf","parameters":[{"name":"cpf","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["clients"]},"post":{"operationId":"ClientController_create","summary":"create client","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientDTO"}}}},"responses":{"201":{"description":""}},"tags":["clients"]}},"/api/v1/products":{"get":{"operationId":"ProductController_getAll","summary":"get products by category","parameters":[{"name":"category","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["products"]},"post":{"operationId":"ProductController_create","summary":"create product","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDTO"}}}},"responses":{"201":{"description":""}},"tags":["products"]}},"/api/v1/products/{id}":{"put":{"operationId":"ProductController_update","summary":"update product by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDTO"}}}},"responses":{"200":{"description":""}},"tags":["products"]},"delete":{"operationId":"ProductController_delete","summary":"delete product by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/api/v1/webhook/createPayment":{"post":{"operationId":"PaymentController_createPaymentWebhook","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentWebhookDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/webhook/receivePayment":{"post":{"operationId":"PaymentController_receivePaymentNotification","parameters":[],"responses":{"201":{"description":""}}}}},"info":{"title":"fastfood soat1","description":"API de pedidos","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"Product":{"type":"object","properties":{"id":{"type":"number"},"code":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"category":{"type":"string"}},"required":["id","code","description","price","category"]},"ItemCart":{"type":"object","properties":{"id":{"type":"number"},"product":{"$ref":"#/components/schemas/Product"},"count":{"type":"number"},"observation":{"type":"string"}},"required":["id","product","count","observation"]},"Cart":{"type":"object","properties":{"id":{"type":"number"},"itens":{"type":"array","items":{"$ref":"#/components/schemas/ItemCart"}}},"required":["id","itens"]},"CreateOrderDTO":{"type":"object","properties":{"cart":{"$ref":"#/components/schemas/Cart"},"observation":{"type":"string"},"totalPrice":{"type":"number"}},"required":["cart","observation","totalPrice"]},"UpdateStatusOrderDTO":{"type":"object","properties":{"status":{"type":"number"}},"required":["status"]},"CreateClientDTO":{"type":"object","properties":{"email":{"type":"string"},"name":{"type":"string"},"cpf":{"type":"string"}},"required":["email","name","cpf"]},"CreateProductDTO":{"type":"object","properties":{"code":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"category":{"type":"string"}},"required":["code","description","price","category"]},"UpdateProductDTO":{"type":"object","properties":{}},"PaymentWebhookDto":{"type":"object","properties":{"amount":{"type":"string"},"orderId":{"type":"string"}},"required":["amount","orderId"]}}}}